styles move to stal
