+++ src/components/Header/Header.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import css from './Header.module.scss';
-import { FiSearch } from "react-icons/fi";
-import styles from '../../styles/Button';
+import { FiSearch } from 'react-icons/fi';
+import stal from '../../pages/stal/Button';
 import login from '../../pages/Login/Login';
 import mypage from '../../pages/MyPage/MyPage';
 import { Link } from 'react-router-dom';
 
-// test 
+// test
 function Header() {
   return (
     <>
@@ -17,26 +17,16 @@ function Header() {
         <div className={css.mypage}>
           <Link to="/mypage">마이페이지</Link>
         </div>
-        <div className={css.interest}>
-          관심상품
-        </div>
-        <div className={css.customer_center}>
-          고객센터
-        </div>
+        <div className={css.interest}>관심상품</div>
+        <div className={css.customer_center}>고객센터</div>
       </div>
       <div className={css.second_border}>
-        <div className={css.kream_log}>
-          DREAM
-        </div>
+        <div className={css.kream_log}>DREAM</div>
         <div className={css.style}>
-          <Link to="/styles">STYLE</Link>
+          <Link to="/stal">STYLE</Link>
         </div>
-        <div className={css.shop}>
-          SHOP
-        </div>
-        <div className={css.about}>
-          ABOUT
-        </div>
+        <div className={css.shop}>SHOP</div>
+        <div className={css.about}>ABOUT</div>
         <div className={css.glass}>
           <FiSearch size="35" />
         </div>
